move types and interfaces to a different file
