+++ src/components/form/Checkbox.tsx
@@ -1,11 +1,8 @@
 import React from 'react';
 import Label from './Label';
+import { CheckboxPropsType } from '../Types';
 
-interface CheckboxProps extends React.ComponentPropsWithoutRef<'input'> {
-  label: string;
-}
-
-const Checkbox = ({ label, ...props }: CheckboxProps) => {
+const Checkbox = ({ label, ...props }: CheckboxPropsType) => {
   return (
     <div className='flex row-reverse flex-end items-center gap-2'>
       <input id={props.id} type='checkbox' {...props} onChange={props.onChange}/>
