+++ src/components/form/Label.tsx
@@ -1,10 +1,7 @@
 import React from 'react'
+import { LabelPropsType } from '../Types'
 
- export interface LabelProps extends React.ComponentPropsWithoutRef <'label'>{
-    label?:string,
-}
-
-const Label = ({ label, ...props }:LabelProps) => {
+const Label = ({ label, ...props }:LabelPropsType) => {
   return (
     <label htmlFor={props.htmlFor} className={`text-xs text-gray-800 dark:text-white`}>{label}</label>
   )
